feature: Expand RandomWordsState to generate infinite scrolling List
RandomWordsState added buildSuggestions() to return ListView implemented to iterate a variable number of row elements incrementing the index and adding a Divider after every pair.
RandomWordsState added buildRows() to return a ListTile containing a WordPair specifying the font size with TextStyle.
RandomWordsState modified build() to return a Scaffold applying material design elements to the ListView returned by _buildSuggestions paired with an AppBar.
MyApp modified the MaterialApp return value to replace the Scaffold set as the home parameter with that returned from RandomWords()